backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pyodbc==5.1.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic>=2.9.2
pydantic[email]>=2.9.2
pydantic-settings>=2.5.2
openai>=2.8.0
jiter>=0.12.0
email-validator==2.1.0
pymupdf>=1.23.0
chardet>=5.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
reportlab>=4.0.0
Pillow>=10.0.0
python-multipart>=0.0.7
httpx>=0.25.0
cryptography>=41.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
